requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
aiosqlite
python-dotenv
streamlit
pandas
scikit-learn
matplotlib
wordcloud